working on keypair generation
