--- net/k8s-cluster/src/config.rs
@@ -1,5 +1,0 @@
-#[derive(Clone, Debug)]
-pub struct SetupConfig<'a> {
-    pub namespace: &'a str,
-    pub num_validators: i32,
-}
